Inline header and footer in page
