--- app/(site)/layout.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function SiteLayout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
-}
